Falta depurar y javadoc

+++ src/main/java/ad/project/backend_wherewatch/domain/repositories/AvailabilityRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AvailabilityRepository extends JpaRepository<Availability, Integer> {
+public interface AvailabilityRepository extends JpaRepository<Availability, Long> {
 }
